requirements.txt
astroid==2.3.1
autopep8==1.4.4
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Hashing==1.1
Flask-JSGlue==0.3.1
Flask-Mail==0.9.1
Flask-MySQL==1.4.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
importlib-metadata==0.23
inflect==3.0.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
mysql-connector-python==8.0.18
mysqlclient==1.4.5
protobuf==3.10.0
pycodestyle==2.5.0
pylint==2.4.2
PyMySQL==0.9.3
random-password-generator==2.1.0
requests==2.22.0
six==1.12.0
sqlacodegen==2.1.0
SQLAlchemy==1.3.10
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
